Squashed commit of the following:
commit ccb11cee2839b8d7791c1907e11c6ca67f85a737
Author: Aldo Gonzalez <totalAldo@gmail.com>
Date:   Fri Apr 24 19:09:59 2026 -0400

    test updates

commit 37da63d3b25109be03ba5844b9f3c66011e3a10b
Author: Aldo Gonzalez <totalAldo@gmail.com>
Date:   Fri Apr 24 18:52:12 2026 -0400

    freebsd fixes

commit af765bd71095599ba7eeebc16dc9095e9e52fcb7
Author: Aldo Gonzalez <totalAldo@gmail.com>
Date:   Fri Apr 24 17:37:51 2026 -0400

    update ci integration tests in github

commit 01d18be7b74bd36bad341c82c73cc19fc98b2d8c
Author: Aldo Gonzalez <totalAldo@gmail.com>
Date:   Fri Apr 24 13:32:56 2026 -0400

    add more tests

commit 64f6d8dd3cab45a2452c2286b17e6699ce63aaf1
Author: Aldo Gonzalez <totalAldo@gmail.com>
Date:   Fri Apr 24 13:12:02 2026 -0400

    update tests

commit 2b76d5afa9a5b399b7f25da7b9380dd841c190f4
Author: Aldo Gonzalez <totalAldo@gmail.com>
Date:   Fri Apr 24 11:03:42 2026 -0400

    text updates

commit b56ed5ca6c9e653ae68f2de06002ab29cccbbd87
Author: Aldo Gonzalez <totalAldo@gmail.com>
Date:   Fri Apr 24 10:02:17 2026 -0400

    test fixes

commit 23291d142d13ff8f8c6cf7e088742d933a839442
Author: Aldo Gonzalez <totalAldo@gmail.com>
Date:   Fri Apr 24 09:12:41 2026 -0400

    checkpoint

+++ src/zxfer_cleanup_child_wrapper.sh
@@ -1,8 +1,40 @@
 #!/bin/sh
+# BSD HEADER START
+# This file is part of zxfer project.
+
+# Copyright (c) 2024-2026 Aldo Gonzalez
+# Copyright (c) 2013-2019 Allan Jude <allanjude@freebsd.org>
+# Copyright (c) 2010,2011 Ivan Nash Dreckman
+# Copyright (c) 2007,2008 Constantin Gonzalez
+# All rights reserved.
+
+# Redistribution and use in source and binary forms, with or without
+# modification, are permitted provided that the following conditions are met:
+
+#     * Redistributions of source code must retain the above copyright notice,
+#       this list of conditions and the following disclaimer.
+#     * Redistributions in binary form must reproduce the above copyright notice,
+#       this list of conditions and the following disclaimer in the documentation
+#       and/or other materials provided with the distribution.
+
+# THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS" AND
+# ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED
+# WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
+# DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT HOLDER OR CONTRIBUTORS BE LIABLE
+# FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL
+# DAMAGES (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR
+# SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER
+# CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY,
+# OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
+# OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
+
+# BSD HEADER END
 # shellcheck shell=sh
 
 zxfer_cleanup_child_wrapper_list_descendants() {
-	l_cleanup_wrapper_snapshot=$(ps -o pid= -o ppid= 2>/dev/null) || return 1
+	l_cleanup_wrapper_snapshot_status=0
+	l_cleanup_wrapper_snapshot=$(ps -o pid= -o ppid= 2>/dev/null) || l_cleanup_wrapper_snapshot_status=$?
+	[ "$l_cleanup_wrapper_snapshot_status" -eq 0 ] || return "$l_cleanup_wrapper_snapshot_status"
 
 	printf '%s\n' "$l_cleanup_wrapper_snapshot" | awk -v root="$$" '
 	{
@@ -34,11 +66,17 @@ zxfer_cleanup_child_wrapper_list_descendants() {
 }
 
 zxfer_cleanup_child_wrapper_abort_descendants() {
-	zxfer_cleanup_child_wrapper_list_descendants |
-		while IFS= read -r l_cleanup_wrapper_pid || [ -n "$l_cleanup_wrapper_pid" ]; do
-			[ -n "$l_cleanup_wrapper_pid" ] || continue
-			kill -s TERM "$l_cleanup_wrapper_pid" 2>/dev/null || :
-		done
+	l_cleanup_wrapper_descendants_status=0
+	l_cleanup_wrapper_descendants=$(zxfer_cleanup_child_wrapper_list_descendants) ||
+		l_cleanup_wrapper_descendants_status=$?
+	[ "$l_cleanup_wrapper_descendants_status" -eq 0 ] || return "$l_cleanup_wrapper_descendants_status"
+
+	while IFS= read -r l_cleanup_wrapper_pid || [ -n "$l_cleanup_wrapper_pid" ]; do
+		[ -n "$l_cleanup_wrapper_pid" ] || continue
+		kill -s TERM "$l_cleanup_wrapper_pid" 2>/dev/null || :
+	done <<-EOF
+		$l_cleanup_wrapper_descendants
+	EOF
 }
 
 zxfer_cleanup_child_wrapper_on_signal() {
@@ -51,15 +89,17 @@ zxfer_cleanup_child_wrapper_main() {
 
 	[ -n "$l_cleanup_wrapper_exec_cmd" ] || return 1
 	trap 'zxfer_cleanup_child_wrapper_on_signal' TERM INT HUP
-	exec 3<&0 || return 1
+	l_cleanup_wrapper_status=0
+	exec 3<&0 || l_cleanup_wrapper_status=$?
+	[ "$l_cleanup_wrapper_status" -eq 0 ] || return "$l_cleanup_wrapper_status"
 
 	# Preserve the wrapper's stdin for background children. Some /bin/sh
 	# implementations reattach asynchronous jobs to /dev/null unless stdin is
 	# duplicated onto a dedicated descriptor before the background launch.
 	sh -c "$l_cleanup_wrapper_exec_cmd" <&3 &
 	l_cleanup_wrapper_child_pid=$!
-	wait "$l_cleanup_wrapper_child_pid"
-	l_cleanup_wrapper_status=$?
+	l_cleanup_wrapper_status=0
+	wait "$l_cleanup_wrapper_child_pid" || l_cleanup_wrapper_status=$?
 	exec 3<&-
 	return "$l_cleanup_wrapper_status"
 }
